requirements.txt
gradio
qrcode
Pillow 